Make price input validation a helper function
